Split lib/analytics/ad-tracking into modules
