+++ client/lib/analytics/ad-tracking/package.json
@@ -1,0 +1,5 @@
+{
+	"sideEffects": [
+		"./setup.js"
+	]
+}
